Add contact, ordered the refenrence from A-Z

+++ Build_docs/Bibtexparser-all.py
@@ -10,10 +10,17 @@ from bibtexparser.bwriter import BibTexWriter
 import os
 import glob
 import subprocess
+import argparse
+
+# Set up argument parser
+parser = argparse.ArgumentParser(description='Fast processing.')
+parser.add_argument('--fast', action='store_true', help='Run the script in fast mode')
+args = parser.parse_args()
 
 # Set BibTeX and parent RST file names
 bib_file_name = "../All.bib"
 parent_rst_name = "By-Cite-Keys.rst"
+parent_rst_html = os.path.splitext(parent_rst_name)[0] + '.html'
 
 # Load BibTeX file
 with open(bib_file_name) as bibtex_file:
@@ -21,35 +28,54 @@ with open(bib_file_name) as bibtex_file:
 
 # Ensure directories for individual entries and bib files exist and empty them
 # before writing new files
-os.makedirs("bib_entries", exist_ok=True)
-os.makedirs("bib_files", exist_ok=True)
-directories = ["bib_entries", "bib_files"]
-for directory in directories:
-    # Get all file paths in the directory
-    files = glob.glob(os.path.join(directory, '*'))
-    # Remove each file
-    for f in files:
-        os.remove(f)
+if not args.fast:
+    os.makedirs("bib_entries", exist_ok=True)
+    os.makedirs("bib_files", exist_ok=True)
+    directories = ["bib_entries", "bib_files"]
+    for directory in directories:
+        # Get all file paths in the directory
+        files = glob.glob(os.path.join(directory, '*'))
+        # Remove each file
+        for f in files:
+            os.remove(f)
 
 # BibTeX writer for creating individual bib files
 writer = BibTexWriter()
 
 # Create/Open the parent RST file
 with open(parent_rst_name, "w") as parent_rst:
-    parent_rst.write("Reference\n")
-    parent_rst.write("=" * len("Reference") + "\n\n")
+    # Add an introductory line or description if needed
+    parent_rst.write("Here is a list of all entries in the SPDEs-Bib database:\n\n")
+    parent_rst.write("References listed by cite keys\n")
+    parent_rst.write("=" * len("References listed by cite keys") + "\n\n")
 
-    for entry in bib_database.entries:
+    # Sort entries by cite_key
+    sorted_entries = sorted(bib_database.entries, key=lambda x: x['ID'].upper())
+
+    current_letter = ''
+    for entry in sorted_entries:
         cite_key = entry["ID"]
+
+        # Check if the first letter has changed
+        first_letter = cite_key[0].upper()
+        if first_letter != current_letter:
+            # Write the section header for this letter
+            current_letter = first_letter
+            parent_rst.write(f"\n{current_letter}\n")
+            parent_rst.write("-" * len(current_letter) + "\n\n")
 
         # Create individual .bib file
         bib_entry_filename = f"bib_files/{cite_key}.bib"
-        with open(bib_entry_filename, "w") as bibfile:
-            db = bibtexparser.bibdatabase.BibDatabase()
-            db.entries = [entry]
-            bibfile.write(writer.write(db))
 
-        subprocess.run(["bibtex-tidy",  "-m", bib_entry_filename])
+        # subprocess.run(["bibtex-tidy",  "-m", bib_entry_filename])
+
+        # Skip certain operations if --fast is set
+        if not args.fast:
+            with open(bib_entry_filename, "w") as bibfile:
+                db = bibtexparser.bibdatabase.BibDatabase()
+                db.entries = [entry]
+                bibfile.write(writer.write(db))
+            subprocess.run(["bibtex-tidy",  "-m", bib_entry_filename])
 
         # Create corresponding RST file
         rst_entry_filename = f"bib_entries/{cite_key}.rst"
@@ -65,22 +91,23 @@ with open(parent_rst_name, "w") as parent_rst:
             with open(bib_entry_filename, "r") as bibfile:
                 for line in bibfile:
                     file.write(f"   {line}")
-            file.write(f"\n`Back to index <../{parent_rst_name}>`_\n")
+            file.write(f"\n`Back to index <../{parent_rst_html}>`_\n")
 
-        # Add entry to the parent RST file
-        parent_rst.write(f"- `{cite_key} <{rst_entry_filename}>`_\n")
+        # Add entry to the parent RST file under the correct section
+        # parent_rst.write(f"- `{cite_key} <{rst_entry_filename}>`_\n")
+        parent_rst.write(f"- :cite:t:`{cite_key}` : `{cite_key} <bib_entries/{cite_key}.html>`_\n")
 
 
-# Create or open the index file
-with open(parent_rst_name, "w") as index_file:
-    # Write the header for the index file
-    index_file.write("List by Citation Keys\n")
-    index_file.write("=======================\n\n")
-
-    # Add an introductory line or description if needed
-    index_file.write("Here is a list of all entries in the SPDEs-Bib database:\n\n")
-
-    # Create links to each entry's page
-    for entry in bib_database.entries:
-        cite_key = entry["ID"]
-        index_file.write(f"- :cite:t:`{cite_key}` : `{cite_key} <bib_entries/{cite_key}.html>`_\n")
+# # Create or open the index file
+# with open(parent_rst_name, "w") as index_file:
+#     # Write the header for the index file
+#     index_file.write("List by Citation Keys\n")
+#     index_file.write("=======================\n\n")
+#
+#     # Add an introductory line or description if needed
+#     index_file.write("Here is a list of all entries in the SPDEs-Bib database:\n\n")
+#
+#     # Create links to each entry's page
+#     for entry in bib_database.entries:
+#         cite_key = entry["ID"]
+#         index_file.write(f"- :cite:t:`{cite_key}` : `{cite_key} <bib_entries/{cite_key}.html>`_\n")
